added jotai and api changes
