+++ .eslintrc
@@ -31,6 +31,7 @@
     "no-unused-vars": "off",
     "react/jsx-uses-react": "off",
     "react/jsx-uses-vars": "warn",
+    "@typescript-eslint/no-unsafe-return": "off",
     "@typescript-eslint/no-explicit-any": "off",
     "@typescript-eslint/no-unsafe-assignment": "off",
     "@typescript-eslint/no-unsafe-argument": "off",
